Move `DefaultDocStore` to `doc-id-default`.
...and, having removed the only class in `data-store-local`, remove that
module.
