+++ local-modules/@bayou/config-server-default/package.json
@@ -4,8 +4,8 @@
     "@bayou/assets-client": "local",
     "@bayou/config-common-default": "local",
     "@bayou/config-server": "local",
-    "@bayou/data-store-local": "local",
     "@bayou/doc-common": "local",
+    "@bayou/doc-id-default": "local",
     "@bayou/env-server": "local",
     "@bayou/file-store-local": "local",
     "@bayou/injecty": "local",
