--- local-modules/@bayou/data-store-local/index.js
@@ -1,7 +1,0 @@
-// Copyright 2016-2019 the Bayou Authors (Dan Bornstein et alia).
-// Licensed AS IS and WITHOUT WARRANTY under the Apache License,
-// Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
-
-import DefaultDocStore from './DefaultDocStore';
-
-export { DefaultDocStore };
