--- local-modules/@bayou/data-store-local/package.json
@@ -1,6 +1,0 @@
-{
-  "dependencies": {
-    "@bayou/config-server": "local",
-    "@bayou/doc-id-default": "local"
- }
-}
