+++ local-modules/@bayou/doc-id-default/README.md
@@ -2,12 +2,11 @@
 ========================
 
 Default definitions of syntax for all the kinds of IDs used by the system. This
-is separate from `config-*-default`, for a couple reasons:
+is separate from `config-*-default`, because non-default configured
+implementations might still want to use these &mdash; quite reasonable &mdash;
+definitions.
 
-* Non-default configured implementations might still want to use these &mdash;
-  quite reasonable &mdash; definitions.
-
-* `data-store-local` needs these definitions. Though it's part of the default
-  configuration, it can be used in other configurations too, and if that's done,
-  we don't want to force the (rest of the) default configuration code to also be
-  pulled in.
+That said, note that the class `DefaultDocStore` is _not_ appropriate for
+production use, because it treats "valid ID syntax" the same as "actually valid
+and existing ID." This is only appropriate for development and testing
+configurations.
