+++ local-modules/@bayou/doc-id-default/index.js
@@ -2,6 +2,7 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
+import DefaultDocStore from './DefaultDocStore';
 import DefaultIdSyntax from './DefaultIdSyntax';
 
-export { DefaultIdSyntax };
+export { DefaultDocStore, DefaultIdSyntax };
